requirements.txt
click==8.1.7
colorama==0.4.6
Flask==2.2.5
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
pillow==10.2.0
Flask_WTF==1.1.1
WTForms==3.0.1
Werkzeug == 2.2.2
